requirements.txt
beautifulsoup4
duckduckgo-search
fastmcp
langchain
langchain-chroma
langchain-community
langchain-core
langchain-experimental
langchain-google-genai
langchain-mcp-adapters
langchain-ollama
langchain-text-splitters
langgraph
langgraph-checkpoint-sqlite
nest-asyncio
pydantic
PyAutoGUI
PyPDF2
python-dotenv
reportlab
requests
streamlit
